Facebook parser stores events into JSON-like structs

+++ src/model/event.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type Event struct {
-    Id uint32
+    Id uint64
 	Title string
 	Desc string
 	Link string
